Add files via upload
Created trash_ent.lua, a ent file that manages the creation of trash entities for the service personnel to clean up, and contains relevant positioning for them to spawn in. Also included are additional changes to support this system within trash_wep (Added functionality to interact and clean up the new ents), and sv_bmrp (added a hook with timer function to actually start spawning these entities after server start).